test: add model_with_logging_metrics fixture to test_dlinear_v2
